requirements.txt
PySide6>=6.6,<6.8
cryptography>=42.0,<45.0
argon2-cffi>=23.1,<25.0
pydantic>=2.6,<3.0
Pillow>=10.0,<12.0
pyinstaller>=6.5,<7.0
pytest>=8.0,<9.0
